requirements.txt
backtrader
alpaca-trade-api
python-dotenv